+++ src/components/Characters/Character.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import { setCharacters } from '../../actions';
-import StyledCharacter from './StyledCharacter';
+import { Character as StyledCharacter } from '../../styles';
 import {
   formatGender,
   calculateHeights,
@@ -89,7 +89,9 @@ const Character = ({ characters, setCharacters }) => {
             <tr>
               <td></td>
               <td></td>
-              <td>{`Total: ${totalHeight}cm (${calculateFeet(totalHeight)}ft/${calculateInches(totalHeight)}in)`}</td>
+              <td>{`Total: ${totalHeight}cm (${calculateFeet(
+                totalHeight
+              )}ft/${calculateInches(totalHeight)}in)`}</td>
             </tr>
           </tbody>
         </table>
