--- src/components/Common/StyledMovieListDropdown.js
@@ -1,67 +1,0 @@
-import styled from 'styled-components';
-
-const MovieListDropdown = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  width: 100%;
-  justify-content: center;
-
-  h1 {
-    width: 100%;
-    text-align: center;
-    font-family: 'Supermercado One', cursive;
-    font-size: 3rem;
-  }
-
-  select {
-    margin: 2rem 0;
-
-    display: block;
-    font-size: 16px;
-    font-family: 'Open Sans', sans-serif;
-    font-weight: 700;
-    color: #444;
-    line-height: 1.3;
-    padding: 0.6em 1.4em 0.5em 0.8em;
-    width: 20%;
-    box-sizing: border-box;
-    border: 1px solid #aaa;
-    box-shadow: 0 1px 0 1px rgba(0, 0, 0, 0.04);
-    border-radius: 0.5em;
-    -moz-appearance: none;
-    -webkit-appearance: none;
-    appearance: none;
-    background-color: #fff;
-    background-image: url('data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%22292.4%22%20height%3D%22292.4%22%3E%3Cpath%20fill%3D%22%23007CB2%22%20d%3D%22M287%2069.4a17.6%2017.6%200%200%200-13-5.4H18.4c-5%200-9.3%201.8-12.9%205.4A17.6%2017.6%200%200%200%200%2082.2c0%205%201.8%209.3%205.4%2012.9l128%20127.9c3.6%203.6%207.8%205.4%2012.8%205.4s9.2-1.8%2012.8-5.4L287%2095c3.5-3.5%205.4-7.8%205.4-12.8%200-5-1.9-9.2-5.5-12.8z%22%2F%3E%3C%2Fsvg%3E'),
-      linear-gradient(to bottom, #ffffff 0%, #e5e5e5 100%);
-    background-repeat: no-repeat, repeat;
-    background-position: right 0.7em top 50%, 0 0;
-    background-size: 0.65em auto, 100%;
-    border-color: black;
-
-    &:hover {
-      /* border-color: #F9F500; */
-      cursor: pointer;
-      -webkit-box-shadow: 0px 0px 3px 0px rgba(0, 0, 0, 0.75);
-      -moz-box-shadow: 0px 0px 3px 0px rgba(0, 0, 0, 0.75);
-      box-shadow: 0px 0px 3px 0px rgba(0, 0, 0, 0.75);
-      transition: all 0.3s ease-in-out;
-    }
-
-    &:-ms-expand {
-      display: none;
-    }
-
-    &:focus {
-      box-shadow: black;
-      /* box-shadow: 0 0 0 3px -moz-mac-focusring; */
-      color: #222;
-      outline: none;
-    }
-
-    option {
-      font-weight: bold;
-    }
-  }
-`;
-export default MovieListDropdown;
